Clean clap_utils
