+++ eth2/utils/clap_utils/Cargo.toml
@@ -12,3 +12,4 @@ hex = "0.3"
 dirs = "2.0"
 types = { path = "../../types" }
 eth2_testnet_config = { path = "../eth2_testnet_config" }
+eth2_ssz = { path = "../ssz" }
