+++ lcli/src/check_deposit_data.rs
@@ -1,11 +1,11 @@
 use clap::ArgMatches;
-use clap_utils::{parse_hex_bytes, parse_required};
+use clap_utils::{parse_required, parse_ssz_required};
 use deposit_contract::{decode_eth1_tx_data, DEPOSIT_DATA_LEN};
 use tree_hash::TreeHash;
 use types::EthSpec;
 
 pub fn run<T: EthSpec>(matches: &ArgMatches) -> Result<(), String> {
-    let rlp_bytes = parse_hex_bytes(matches, "deposit-data")?;
+    let rlp_bytes = parse_ssz_required::<Vec<u8>>(matches, "deposit-data")?;
     let amount = parse_required(matches, "deposit-amount")?;
 
     if rlp_bytes.len() != DEPOSIT_DATA_LEN {
